WIP: export fixes, mro issues
