+++ src/spyglass/common/common_dandi.py
@@ -1,7 +1,6 @@
 import os
 import shutil
 from pathlib import Path
-from typing import Optional
 
 import datajoint as dj
 import fsspec
